switched fields representing currency (for example transaction amount) to use decimal type instead of double to prevent floating point errors

+++ CommerceBankWebApp/Areas/Identity/Pages/Account/Manage/AddNotificationRule.cshtml.cs
@@ -45,7 +45,7 @@ namespace CommerceBankWebApp.Areas.Identity.Pages.Account.Manage
 
             [Required]
             [Display(Name = "Threshold")]
-            public double Threshold { get; set; }
+            public decimal Threshold { get; set; }
         }
 
         private async Task LoadAsync(IdentityUser user)
